+++ Documentation/Elements/Interaction/TCResizable
