+++ Documentation/Elements/Interaction/TC_box2_Resizable
